server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "aiomysql>=0.2.0",
    "alembic>=1.18.4",
    "fastapi>=0.136.1",
    "greenlet>=3.5.0",
    "openai>=2.33.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.49",
    "tenacity>=9.1.4",
    "uvicorn>=0.46.0",
]
